+++ CHANGES.rst
@@ -5,6 +5,9 @@ Changelog
 1.2.4 (unreleased)
 ------------------
 
+- WEB-3301: Adapt "state" column on citizen dashboard to display only the current state
+  [mpeeters]
+
 - WEB-3301: Fix "online" column value for citizens by adding a new config parameter that define which workflow states are egal to "online"
   [mpeeters]